move notifier_protocol.py
