--- hub/elastic_sync/__init__.py
@@ -1,2 +1,0 @@
-from hub.elastic_sync.search import SearchIndex
-from hub.elastic_sync.notifier_protocol import ElasticNotifierClientProtocol
